Refine Tooltip enhancement implementation per @TejasQ's review
