+++ packages/components/src/Tooltip/Tooltip.tsx
@@ -1,8 +1,10 @@
 import * as React from "react"
 import styled from "react-emotion"
+
 import { OperationalStyleConstants } from "../utils/constants"
 import colorCalculator from "tinycolor2"
 import { Css } from "../types"
+import Container, { Position } from "./Tooltip.Container"
 
 /**
  * In order to allow for tooltips that have a sensible max-width that adjusts its width for shorter text,
@@ -11,6 +13,7 @@ import { Css } from "../types"
  * The actual tooltip is rendered with this information extracted from the DOM node.
  */
 
+/** @todo Type this more strongly with discriminated unions to prevent unexpected behavior for impossible prop combinations such as `<Tooltip right bottom />`. */
 export interface Props {
   className?: string
   children?: React.ReactNode
@@ -34,142 +37,6 @@ export interface State {
   bbRight: number
   singleLineTextWidth: number
 }
-
-type Position = "top" | "left" | "right" | "bottom"
-
-const containerPositionStyles = (position: Position): {} => {
-  switch (position) {
-    case "top":
-      return {
-        left: "50%",
-        top: -6,
-        transform: "translate3d(-50%, -100%, 0)",
-      }
-    case "bottom":
-      return {
-        left: "50%",
-        top: "100%",
-        transform: "translate3d(-50%, 6px, 0)",
-      }
-    case "left":
-      return {
-        top: "50%",
-        left: -6,
-        transform: "translate3d(-100%, -50%, 0)",
-      }
-    case "right":
-      return {
-        top: "50%",
-        right: -6,
-        transform: "translate3d(100%, -50%, 0)",
-      }
-  }
-}
-
-const pointerTrianglePositionStyles = (position: Position, backgroundColor: string): {} => {
-  switch (position) {
-    case "top":
-      return {
-        bottom: -4,
-        left: `calc(50% - 6px)`,
-        borderLeft: "6px solid transparent",
-        borderRight: "6px solid transparent",
-        borderTop: `6px solid ${backgroundColor}`,
-      }
-    case "bottom":
-      return {
-        top: -4,
-        left: `calc(50% - 6px)`,
-        borderLeft: "6px solid transparent",
-        borderRight: "6px solid transparent",
-        borderBottom: `6px solid ${backgroundColor}`,
-      }
-    case "left":
-      return {
-        right: -4,
-        top: `calc(50% - 6px)`,
-        borderTop: "6px solid transparent",
-        borderBottom: "6px solid transparent",
-        borderLeft: `6px solid ${backgroundColor}`,
-      }
-    case "right":
-      return {
-        left: -4,
-        top: `calc(50% - 6px)`,
-        borderTop: "6px solid transparent",
-        borderBottom: "6px solid transparent",
-        borderRight: `6px solid ${backgroundColor}`,
-      }
-  }
-}
-
-const Container = styled("div")(
-  ({
-    position,
-    offScreenWidthTest,
-    singleLineTextWidth,
-    theme,
-  }: {
-    position: Position
-    offScreenWidthTest?: boolean
-    singleLineTextWidth: number
-    theme?: OperationalStyleConstants
-  }) => {
-    const backgroundColor = colorCalculator(theme.color.black)
-      .setAlpha(0.9)
-      .toString()
-    return {
-      backgroundColor,
-      label: "tooltip",
-      color: theme.color.white,
-      position: "absolute",
-      zIndex: theme.zIndex.tooltip,
-      borderRadius: 2,
-      boxShadow: "0 2px 6px rgba(0, 0, 0, .15)",
-      "& > p": {
-        fontSize: 11,
-        lineHeight: 1.3,
-        margin: 0,
-        padding: "2px 6px",
-        textAlign: "center",
-      },
-      ...(offScreenWidthTest
-        ? {
-            width: "fit-content",
-            whiteSpace: "nowrap",
-            position: "fixed",
-            opacity: 0.01,
-            top: -200,
-            left: -200,
-          }
-        : {
-            // If there was an issue determining singleLineTextWidth, default to the 150px width
-            // Otherwise, honor the single line text width unless greater than 150px.
-            width: singleLineTextWidth === 0 ? 150 : Math.min(singleLineTextWidth + 4, 150),
-          }),
-      ...containerPositionStyles(position),
-      // This pseudo-element extends the clickable area of a tooltip extending enough to disappear as the mouse moves over to the caret.
-      "&::after": {
-        content: "''",
-        position: "absolute",
-        top: 0,
-        left: -32,
-        display: "block",
-        width: -32,
-        height: "100%",
-      },
-      // They say behind every great tooltip is a great caret.
-      "&::before": {
-        content: "''",
-        position: "absolute",
-        zIndex: theme.zIndex.tooltip,
-        width: 0,
-        height: 0,
-        ...pointerTrianglePositionStyles(position, backgroundColor),
-      },
-    }
-  },
-)
 
 class Tooltip extends React.Component<Props, State> {
   state = {
@@ -202,7 +69,7 @@ class Tooltip extends React.Component<Props, State> {
     this.setDomProperties()
   }
 
-  render() {
+  getPosition() {
     let position: Position = "right"
 
     if (this.props.left) {
@@ -221,6 +88,12 @@ class Tooltip extends React.Component<Props, State> {
       position = "top"
     }
 
+    return position
+  }
+
+  getDisplayPosition() {
+    let position: Position = this.getPosition()
+
     if (this.props.smart) {
       /** @todo implement bounding box checks for right- and bottom-placed tooltips.
        * This should be easier once the OperationalUI provides window dimensions in context.
@@ -234,9 +107,17 @@ class Tooltip extends React.Component<Props, State> {
       }
     }
 
+    return position
+  }
+
+  render() {
+    const displayPosition = this.getDisplayPosition()
+
     return (
       <>
-        {/* Test node rendered to determine how wide the text is if it were written in a single line. */}
+        {/* Test node rendered to determine how wide the text is if it were written in a single line.
+          * Note that the position is set arbitrarily since it does not influence text width.
+          */}
         <Container
           position="bottom"
           offScreenWidthTest
@@ -250,7 +131,7 @@ class Tooltip extends React.Component<Props, State> {
         </Container>
         <Container
           singleLineTextWidth={this.state.singleLineTextWidth}
-          position={position}
+          position={displayPosition}
           innerRef={node => {
             this.containerNode = node
           }}
